# Downloads/Stege_freddy/my-fullstack-app/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Learning Platform API",
    "version": "1.0.0",
    "description": "API documentation for the Learning Platform backend."
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register user (admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "roleId": { "type": "integer" }
                },
                "required": ["username", "email", "password", "roleId"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/projects": {
      "post": {
        "summary": "Create project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "deadline": { "type": "string", "format": "date-time" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Project created" }
        }
      }
    },
    "/api/tokens/award": {
      "post": {
        "summary": "Award tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "integer" },
                  "type": { "type": "string" },
                  "value": { "type": "integer" }
                },
                "required": ["userId", "type", "value"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Tokens awarded" }
        }
      }
    },
    "/api/notifications/send": {
      "post": {
        "summary": "Send notification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "integer" },
                  "message": { "type": "string" },
                  "type": { "type": "string" },
                  "scheduledFor": { "type": "string", "format": "date-time" }
                },
                "required": ["userId", "message", "type"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Notification sent" }
        }
      }
    }
  }
}